stravax/chatbot/requirements.txt
google-generativeai
python-dotenv
elevenlabs
sounddevice
soundfile
keyboard
